\name{nparcompGrouping}
\alias{nparcompGrouping}
\title{nparcompGrouping}
\usage{

nparcompGrouping(nparcomGrouping = nparcomp_results)

}

\description{

nparcompGrouping(nparcomGrouping, pval=0.05, capital_latter = F)

pval : a p value. default is 0.05

capital_letter: Group letter print to a capital.Default is FALSE

}

\examples{

}
